refactor: moved RentForm to own File
